requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
